Refactor du code
